encoder_to_odom: unbreak build
Turns out the directory /home/nyx/dev when added to the ROS_PACKAGE_PATH
makes rosdep really slow (hanging?).

Also, for every directory in ROS_PACKAGE_PATH it stats all files in that
dir to determine if they are directories and repeats (in addition to
trying to read manifest.xml and rosdep_nodirs [or something like that]).